requirements.txt
torch==2.3.1
torchvision==0.18.1
boto3==1.20.39
omegaconf==2.1.1
transformers==4.50.1
accelerate==1.2.1
peft==0.14.0
deepspeed==0.12.6
pandas
numpy==1.26.4
tqdm
loguru
click
tensorboard
bitsandbytes
sentencepiece
evaluate
scikit-learn
nltk
dacite
torchmetrics
intervaltree
python-dotenv
clearml
